Remove steer angle conversion.
This is basically the same as https://github.com/lgsvl/simulator/pull/1372, except that the target wheel angle is left unmodified (assumed to be in range `[-1, 1]`) instead of treating it as radians.
While this is in conflict with the comment on the `lgsvl_msgs` repo, it seems to match the implementation of the `lgsvl_interface` in `AutowareAuto`.